big decimal for volumes and values added access modifier changed object relations for user-subscription and beer table

+++ src/main/java/by/alexeysavchic/beer_pet_project/entity/BeerCharacteristics.java
@@ -1,6 +1,6 @@
 package by.alexeysavchic.beer_pet_project.entity;
 
-import by.alexeysavchic.beer_pet_project.entity.id.BeerCharacteristicsId;
+import by.alexeysavchic.beer_pet_project.entity.ids.BeerCharacteristicsId;
 import by.alexeysavchic.beer_pet_project.entity.enums.BeerCharacteristic;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -15,6 +15,7 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -36,5 +37,5 @@ public class BeerCharacteristics
 
     @Column(name = "value")
     @Positive
-    private Double value;
+    private BigDecimal value;
 }
